Moved bill.js, cleaned up index.js, added separate route for subscriber dashboard and fixed small bug
