+++ src/App.js
@@ -6,6 +6,7 @@ import Onboarding from './screens/onboarding/Onboarding';
 import Login from './screens/auth/Login';
 import SignUp from './screens/auth/SignUp';
 import SubscriberOwnerDashboard from './screens/general/SubscriberOwnerDashboard';
+import GeneralOwnerDashboard from './screens/general/GeneralOwnerDashboard';
 
 function App() {
   return (
@@ -17,8 +18,8 @@ function App() {
           <Route exact path="/" component={Login} />
           <Route path="/signup" component={SignUp} />
           <Route path="/onboarding" component={Onboarding} />
-          {/* <Route path="/dashboard" component={GeneralOwnerDashboard}/>  */}
-          <Route path="/dashboard" component={SubscriberOwnerDashboard} />
+          <Route path="/dashboard" component={GeneralOwnerDashboard} />
+          <Route path="/subdashboard" component={SubscriberOwnerDashboard} />
           <Route>Not Found - 404</Route>
         </Switch>
       </div>
