+++ src/index.js
@@ -1,17 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './styles/index.css';
-
 import App from './App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
-
-// REEACT ROUTER
-// import RoutingApp from './RoutingApp';
-// import { BrowserRouter } from 'react-router-dom';
-
-// ReactDOM.render((
-//   <BrowserRouter>
-//     <RoutingApp />
-//   </BrowserRouter>
-// ), document.getElementById('root'));
